+++ src/napari_metadata/_axis_type.py
@@ -1,0 +1,15 @@
+from enum import Enum, auto
+from typing import List
+
+
+class AxisType(Enum):
+    SPACE = auto()
+    TIME = auto()
+    CHANNEL = auto()
+
+    def __str__(self) -> str:
+        return self.name.lower()
+
+    @classmethod
+    def names(cls) -> List[str]:
+        return [str(m) for m in cls]
