+++ src/napari_metadata/_time_units.py
@@ -1,0 +1,16 @@
+from enum import Enum, auto
+from typing import List
+
+
+class TimeUnits(Enum):
+    NANOSECONDS = auto()
+    MICROSECONDS = auto()
+    MILLISECONDS = auto()
+    SECONDS = auto()
+
+    def __str__(self) -> str:
+        return self.name.lower()
+
+    @classmethod
+    def names(cls) -> List[str]:
+        return [str(m) for m in cls]
